Fix: 문제번호 입력시 제목 안나가는 issue 해결

+++ src/main/java/com/example/baekersolved/domain/SolvedApiService.java
@@ -126,7 +126,7 @@ public class SolvedApiService {
         return studyRule.getData();
     }
 
-    public String getSolvedSubject(Long problemId) throws Exception{
+    public String getSolvedSubject(int problemId) throws Exception{
         return solvedApiManager.getSubject(problemId);
     }
